refactor info import into service
